Changed example to reflect the current form of data.
The major change is, that the Person cannot be referenced directly
from the Image with hasAgent any more. In order to convert all data, I
had to create anonymous nodes describing an Agent' Membership in a
Media, i.e., a Movie is connected to the ProductionMember instance
with has Member. The ProductionMember subclasses describe different
role a member can have in production, and the actual person is
connected to the ProductionMember with hasAgent.

In short, the link from an Image to a Person has to go via
ProductionMember which can be treated anonymousy like in this example.