Elonet person data (#7)
* Collect person data from Elonet www pages

* Document elonet-person-data.py

* Downloads data from Elonet and creates a Turtle

The date-data is not parsed as dates, yet. The data from Elonet is so
bad, that parsing is difficult. The script is so slow, that re-running
it again and again is too slow a method, and parsing the date strings
must be done with the data inserted in triple store.

* minor fix to script description

* Script to clean the person data dates

Co-authored-by: Harri <harri@epinurmo.kotikoti>